Material Setup Improvements
Currently, environment maps are not being used correctly, so my approach is to actually sample them using a Reflection vector and sending them to the Emission Color input of the BSDF.

Additionally, when a material's base color map is implied to be opaque, the alpha channel will be disconnected, this way, Blender will treat them as truly opaque. Currently this only works when importing models with the CTXR textures, which have file names that contain certain keywords that indicate their intended blending mode.