Merge Material Slots import setting
Instead of creating a new redundant Material and a separate Material Slot for every Vertex Group, 'Merge Similar Material Slots' can be enabled so that importers will keep track of duplicate material settings and collapse them all under a single material and slot when constructing the mesh.

A material is considered to be the same as another when they share the same texture maps and flags.

Additionally, materials are now named after their Base Color Texture name.